feat: oauth2와 로컬 로직 동일하게 변경

+++ backend/service/user/src/main/java/com/mung/mungtique/user/adaptor/in/web/OAuth2Controller.java
@@ -2,14 +2,18 @@ package com.mung.mungtique.user.adaptor.in.web;
 
 import com.mung.mungtique.user.application.port.in.TokenService;
 import com.mung.mungtique.user.infrastructure.jwt.JwtUtil;
-import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.security.oauth2.jwt.JwtException;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Arrays;
 import java.util.Optional;
@@ -27,41 +31,46 @@ public class OAuth2Controller {
     public ResponseEntity<String> logout(HttpServletRequest request, HttpServletResponse response) {
         log.info("OAuth2 로그아웃 요청됨");
 
-        // 토큰 레디스에서 삭제
         String refreshToken = Optional.ofNullable(request.getCookies())
                 .flatMap(cookies -> Arrays.stream(cookies)
-                        .filter(cookie -> "Oauth2-Access-Token".equals(cookie.getName()))
+                        .filter(cookie -> "refresh".equals(cookie.getName()))
                         .map(Cookie::getValue)
                         .findFirst()
                 ).orElse(null);
-        tokenService.deleteRefreshToken(refreshToken);
+
+        // 토큰 레디스에서 삭제
+        if (refreshToken != null) {
+            tokenService.deleteRefreshToken(refreshToken);
+            log.info("Deleted refresh token from Redis: {}", refreshToken);
+        } else {
+            log.warn("Refresh token not found in cookies.");
+        }
 
         // 쿠키 모두 삭제
-        Arrays.asList("Oauth2-Access-Token", "JSESSIONID")
+        Arrays.asList("refresh", "JSESSIONID")
                 .forEach(name -> deleteCookie(response, name));
 
         return ResponseEntity.ok("OAuth2 logout complete");
     }
 
     @GetMapping("/oauth2/me")
-    public ResponseEntity<String> getUserInfo(@CookieValue(name = "Oauth2-Access-Token", required = false) String accessToken,
-                                                                                            HttpServletResponse response) {
-        log.info("check Oauth2 access token : {}", accessToken);
+    public ResponseEntity<String> getUserInfo(HttpServletRequest request, HttpServletResponse response) {
+        String accessToken = request.getHeader("Authorization");
 
         if (accessToken == null) return ResponseEntity.ok("no access token");
 
+        accessToken = accessToken.substring(7);
+
         try {
-            String userId = jwtUtil.extractSubject(accessToken);
-            log.info("인증된 userId: {}", userId);
-            return ResponseEntity.ok(userId);
-        } catch (ExpiredJwtException ex) {
-            log.warn("Access Token 만료됨: {}", ex.getMessage());
-            Arrays.asList("Oauth2-Access-Token", "JSESSIONID")
-                    .forEach(name -> deleteCookie(response, name));
-            return ResponseEntity.ok("expired");
+            Long userId = jwtUtil.extractUserId(accessToken);
+            log.info("oauth2 getUserInfo userId: {}", userId);
+            return ResponseEntity.ok(userId.toString());
+        }  catch (JwtException ex) {
+            log.error("JWT 오류: {}", ex.getMessage());
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid or expired token");
         } catch (Exception ex) {
             log.error("서버 오류 발생: {}", ex.getMessage());
-            return ResponseEntity.ok("error");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Server error");
         }
     }
